# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profile-verifications": {
      "post": {
        "operationId": "ProfileVerificationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "연락처 인증 발송 — 현재 비밀번호 재확인 + 채널(email/sms) + 대상. 응답은 masked만.",
        "tags": [
          "profile-verifications"
        ]
      }
    },
    "/api/profile-verifications/{id}/confirm": {
      "post": {
        "operationId": "ProfileVerificationsController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmProfileVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "인증 코드 확인 — 실패 5회 잠금(400 일반화 메시지), 성공 시 verified.",
        "tags": [
          "profile-verifications"
        ]
      }
    },
    "/api/profile-verifications/{id}/resend": {
      "post": {
        "operationId": "ProfileVerificationsController_resend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "인증 코드 재전송 — cooldown 60초·최대 5회, 만료 10분 갱신.",
        "tags": [
          "profile-verifications"
        ]
      }
    },
    "/api/profile-change-requests/mine": {
      "get": {
        "operationId": "ProfileChangeRequestsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "내 프로필 변경 요청 목록. [전 직원]",
        "tags": [
          "profile-change-requests"
        ]
      }
    },
    "/api/profile-change-requests": {
      "post": {
        "operationId": "ProfileChangeRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileChangeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "프로필 변경 요청 생성. 이름·전화·국가·타임존만 허용. [전 직원]",
        "tags": [
          "profile-change-requests"
        ]
      },
      "get": {
        "operationId": "ProfileChangeRequestsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "전체 프로필 변경 요청 목록. [관리자]",
        "tags": [
          "profile-change-requests"
        ]
      }
    },
    "/api/profile-change-requests/{id}": {
      "get": {
        "operationId": "ProfileChangeRequestsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "프로필 변경 요청 상세. 요청자 본인 또는 관리자.",
        "tags": [
          "profile-change-requests"
        ]
      }
    },
    "/api/profile-change-requests/{id}/approve": {
      "post": {
        "operationId": "ProfileChangeRequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "프로필 변경 요청 승인. 본인 승인 불가. [관리자]",
        "tags": [
          "profile-change-requests"
        ]
      }
    },
    "/api/profile-change-requests/{id}/reject": {
      "post": {
        "operationId": "ProfileChangeRequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProfileChangeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileChangeRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "프로필 변경 요청 반려. 사유 필수, 본인 반려 불가. [관리자]",
        "tags": [
          "profile-change-requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "TACO ERP API",
    "description": "TnAcademy 백오피스 운영 API. Postgres 영속 저장소와 역할 기반 접근 제어를 사용하는 실제 서비스 계약.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "현재 호스트"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "CreateProfileVerificationDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 72,
            "description": "현재 비밀번호(재확인) — 저장·로그 금지"
          },
          "channel": {
            "enum": [
              "email",
              "sms"
            ],
            "type": "string",
            "description": "인증 채널"
          },
          "target": {
            "type": "string",
            "minLength": 3,
            "maxLength": 320,
            "description": "변경할 연락처(이메일 주소 또는 휴대전화 — E.164/국내형)"
          }
        },
        "required": [
          "currentPassword",
          "channel",
          "target"
        ]
      },
      "ProfileVerificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "channel": {
            "enum": [
              "email",
              "sms"
            ],
            "type": "string"
          },
          "maskedTarget": {
            "type": "string",
            "description": "마스킹된 대상(예: pa***@t***.test, +82*****678)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "consumed",
              "expired",
              "locked"
            ]
          },
          "expiresAt": {
            "type": "string",
            "description": "만료 시각(ISO)"
          },
          "resendAvailableAt": {
            "type": "string",
            "description": "재전송 가능 시각(ISO)"
          },
          "attemptsLeft": {
            "type": "number",
            "description": "남은 확인 시도 횟수"
          }
        },
        "required": [
          "id",
          "channel",
          "maskedTarget",
          "status",
          "expiresAt",
          "resendAvailableAt"
        ]
      },
      "ConfirmProfileVerificationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 4,
            "maxLength": 10,
            "pattern": "^\\d+$",
            "description": "수신한 인증 코드(6자리 숫자)"
          }
        },
        "required": [
          "code"
        ]
      },
      "ProfileChangeRequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "requesterId": {
            "type": "number"
          },
          "baseProfileVersion": {
            "type": "number",
            "minimum": 1
          },
          "beforeValues": {
            "type": "object",
            "additionalProperties": true
          },
          "requestedChanges": {
            "type": "object",
            "additionalProperties": true
          },
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "decidedBy": {
            "type": "number",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "appliedProfileVersion": {
            "type": "number",
            "nullable": true,
            "minimum": 2
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "requesterId",
          "baseProfileVersion",
          "beforeValues",
          "requestedChanges",
          "reason",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProfileChangeRequestDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 72,
            "description": "현재 비밀번호(재확인)"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "박지훈"
          },
          "webId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "jihoon_park",
            "description": "변경할 아이디(대표 승인 후 적용·재로그인 필요)"
          },
          "email": {
            "type": "string",
            "maxLength": 320,
            "format": "email",
            "example": "new@tnacademy.test",
            "description": "변경할 이메일(사전 인증 필수)"
          },
          "verificationChallengeId": {
            "type": "number",
            "description": "연락처(email/phone) 변경 시 소비할 verified challenge id"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20,
            "example": "+82-10-1234-5678"
          },
          "countryCode": {
            "type": "string",
            "nullable": true,
            "pattern": "^[A-Za-z][A-Za-z0-9-]{1,7}$",
            "example": "KR",
            "description": "국가/권역 코드(예: KR, US-W)"
          },
          "timeZone": {
            "type": "string",
            "nullable": true,
            "maxLength": 64,
            "example": "Asia/Seoul"
          },
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500,
            "example": "업무 연락처가 변경되었습니다."
          }
        },
        "required": [
          "currentPassword",
          "reason"
        ]
      },
      "RejectProfileChangeRequestDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500,
            "example": "전화번호 확인 후 다시 요청해 주세요."
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "access_token": []
    }
  ]
}